Change to trait-based approach
